// src/app/models/patient.ts
export interface Patient {
    documentNumber: number;
    documentType: string;
    name1: string;
    name2: string;
    lastname1: string;
    lastname2: string;
    birthdate: string;
    gender: string;
    sex: string;
    address: string;
    phoneNumber: string;
    planTypeId: number; 
}